change name of else_if_used to collapsible_else_if and make it actually work
