+++ crates/ruff/resources/test/fixtures/pylint/collapsible_else_if.py
