+++ crates/ruff/src/rules/pylint/rules/mod.rs
@@ -2,9 +2,9 @@ pub use await_outside_async::{await_outside_async, AwaitOutsideAsync};
 pub use bad_str_strip_call::{bad_str_strip_call, BadStrStripCall};
 pub use bad_string_format_type::{bad_string_format_type, BadStringFormatType};
 pub use bidirectional_unicode::{bidirectional_unicode, BidirectionalUnicode};
+pub use collapsible_else_if::{collapsible_else_if, CollapsibleElseIf};
 pub use comparison_of_constant::{comparison_of_constant, ComparisonOfConstant};
 pub use consider_using_sys_exit::{consider_using_sys_exit, ConsiderUsingSysExit};
-pub use else_if_used::{else_if_used, ElseIfUsed};
 pub use global_variable_not_assigned::GlobalVariableNotAssigned;
 pub use invalid_all_format::{invalid_all_format, InvalidAllFormat};
 pub use invalid_all_object::{invalid_all_object, InvalidAllObject};
@@ -34,9 +34,9 @@ mod await_outside_async;
 mod bad_str_strip_call;
 mod bad_string_format_type;
 mod bidirectional_unicode;
+mod collapsible_else_if;
 mod comparison_of_constant;
 mod consider_using_sys_exit;
-mod else_if_used;
 mod global_variable_not_assigned;
 mod invalid_all_format;
 mod invalid_all_object;
